[HEAD-415] Use setup approach for ruby-app zones upgrades.
ruby-app zones will use a backup of the app-data filesystem when exists instead of create a new one.
build upgrade will just checkout the ruby-app zones instead of download products from app-release builder (no longer required)

+++ zones/capi/95-zone-datasets.sh
@@ -23,7 +23,11 @@ for app in capi dnsapi; do
     mkdir -p /opt/smartdc/$app-data/log
     mkdir -p /opt/smartdc/$app-data/tmp/pids
     # Remove and symlink directories:
-    mv /opt/smartdc/$app/config /opt/smartdc/$app-data/config
+    if [[ ! -n ${KEEP_DATA_DATASET} ]]; then
+      mv /opt/smartdc/$app/config /opt/smartdc/$app-data/config
+    else
+      rm -Rf /opt/smartdc/$app/config
+    fi
     rm -Rf /opt/smartdc/$app/log
     rm -Rf /opt/smartdc/$app/tmp
     rm -Rf /opt/smartdc/$app/config
@@ -31,7 +35,9 @@ for app in capi dnsapi; do
     ln -s /opt/smartdc/$app-data/tmp /opt/smartdc/$app/tmp
     ln -s /opt/smartdc/$app-data/config /opt/smartdc/$app/config
     # Save REVISION:
-    echo "${REVISION}">/opt/smartdc/$app-data/REVISION
+    if [[ ! -n ${KEEP_DATA_DATASET} ]]; then
+      echo "${REVISION}">/opt/smartdc/$app-data/REVISION
+    fi
     echo "${REVISION}">/opt/smartdc/$app/REVISION
     # Save VERSION (Updates based on this):
     APP_VERSION=$(/opt/local/bin/git describe --tags)
